# docs/openapi/schedule.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
...
  "paths" : {
    "/schedule/items/{id}" : {
      "get" : {
        "tags" : [ "Schedule Items" ],
        "summary" : "Получить шаблон по ID",
        "operationId" : "getScheduleItem",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Шаблон найден",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Шаблон не найден",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации запроса",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Доступ запрещён",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Schedule Items" ],
        "summary" : "Полное обновление шаблона (PUT, HEADMAN/ADMIN)",
        "operationId" : "updateScheduleItem",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateScheduleItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Шаблон обновлён",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Нет прав доступа",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Шаблон не найден",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "503" : {
            "description" : "Academic Service недоступен",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Schedule Items" ],
        "summary" : "Деактивировать шаблон расписания (HEADMAN/ADMIN)",
        "operationId" : "deleteScheduleItem",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Шаблон деактивирован"
          },
          "403" : {
            "description" : "Нет прав доступа",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Шаблон не найден",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации запроса",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/schedule/items" : {
      "get" : {
        "tags" : [ "Schedule Items" ],
        "summary" : "Список шаблонов расписания для группы и семестра",
        "operationId" : "listScheduleItems",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "semesterId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        }, {
          "name" : "assembler",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PagedResourcesAssemblerScheduleItemResponse"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Список шаблонов",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedModelEntityModelScheduleItemResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации запроса",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Доступ запрещён",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Ресурс не найден",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт данных",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Schedule Items" ],
        "summary" : "Создать шаблон расписания (HEADMAN/ADMIN)",
        "operationId" : "createScheduleItem",
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateScheduleItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Шаблон создан",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Ошибка валидации",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Нет прав доступа",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Конфликт assignment fence или idempotency replay",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "503" : {
            "description" : "Academic Service недоступен",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Требуется аутентификация",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Ресурс не найден",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "Превышен лимит запросов",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Внутренняя ошибка сервера",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UpdateScheduleItemRequest" : {
        "type" : "object",
        "description" : "Запрос на полное обновление (PUT) шаблона пары; groupId/semesterId неизменяемы (D-09)",
        "properties" : {
          "subjectId" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "ID предмета",
            "example" : 42
          },
          "dayOfWeek" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "День недели: 1 — понедельник, 7 — воскресенье",
            "example" : 1,
            "maximum" : 7,
            "minimum" : 1
          },
          "lessonNumber" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Номер пары в дне (1..8)",
            "example" : 2,
            "maximum" : 8,
            "minimum" : 1
          },
          "startTime" : {
            "type" : "string",
            "description" : "Время начала пары",
            "example" : "10:30:00"
          },
          "endTime" : {
            "type" : "string",
            "description" : "Время окончания пары",
            "example" : "12:00:00"
          },
          "weekType" : {
            "type" : "string",
            "description" : "Тип недели: ODD — 1-я (ISO чётная), EVEN — 2-я (ISO нечётная)",
            "enum" : [ "ALL", "ODD", "EVEN" ],
            "example" : "EVEN"
          },
          "room" : {
            "type" : "string",
            "description" : "Номер аудитории",
            "example" : "3-405",
            "maxLength" : 64,
            "minLength" : 0
          }
        },
        "required" : [ "dayOfWeek", "endTime", "lessonNumber", "startTime", "subjectId", "weekType" ]
      },
      "EntityModelScheduleItemResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "assignmentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "subjectId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "semesterId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dayOfWeek" : {
            "type" : "integer",
            "format" : "int32"
          },
          "lessonNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "startTime" : {
            "type" : "string"
          },
          "endTime" : {
            "type" : "string"
          },
          "weekType" : {
            "type" : "string",
            "enum" : [ "ALL", "ODD", "EVEN" ]
          },
          "room" : {
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "generatedCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "generatedFrom" : {
            "type" : "string",
            "format" : "date"
          },
          "generatedUntil" : {
            "type" : "string",
            "format" : "date"
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          }
        }
      },
...
      "CreateScheduleItemRequest" : {
        "type" : "object",
        "description" : "Запрос на создание шаблона пары в расписании (повторяющийся слот)",
        "properties" : {
          "assignmentId" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "ID авторитетного назначения преподаватель–предмет–группа",
            "example" : 501,
            "minimum" : 1
          },
          "groupId" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "ID группы",
            "example" : 10
          },
          "subjectId" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "ID предмета",
            "example" : 42
          },
          "semesterId" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "ID семестра",
            "example" : 3
          },
          "dayOfWeek" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "День недели: 1 — понедельник, 7 — воскресенье",
            "example" : 1,
            "maximum" : 7,
            "minimum" : 1
          },
          "lessonNumber" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Номер пары в дне (1..8)",
            "example" : 2,
            "maximum" : 8,
            "minimum" : 1
          },
          "startTime" : {
            "type" : "string",
            "description" : "Время начала пары",
            "example" : "10:30:00"
          },
          "endTime" : {
            "type" : "string",
            "description" : "Время окончания пары",
            "example" : "12:00:00"
          },
          "weekType" : {
            "type" : "string",
            "description" : "Тип недели: ODD — 1-я (ISO чётная), EVEN — 2-я (ISO нечётная)",
            "enum" : [ "ALL", "ODD", "EVEN" ],
            "example" : "ODD"
          },
          "room" : {
            "type" : "string",
            "description" : "Номер аудитории",
            "example" : "3-405",
            "maxLength" : 64,
            "minLength" : 0
          }
        },
        "required" : [ "assignmentId", "dayOfWeek", "endTime", "groupId", "lessonNumber", "semesterId", "startTime", "subjectId", "weekType" ]
      },
...
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "sort" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "PagedResourcesAssemblerScheduleItemResponse" : {
        "type" : "object",
        "properties" : {
          "forceFirstAndLastRels" : {
            "type" : "boolean",
            "writeOnly" : true
          }
        }
      },
      "PageMetadata" : {
        "type" : "object",
        "properties" : {
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "PagedModelEntityModelScheduleItemResponse" : {
        "type" : "object",
        "properties" : {
          "_embedded" : {
            "type" : "object",
            "properties" : {
              "scheduleItemResponseList" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/EntityModelScheduleItemResponse"
                }
              }
            }
          },
          "_links" : {
            "$ref" : "#/components/schemas/Links"
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        }
      },
...
      "Link" : {
        "type" : "object",
        "properties" : {
          "href" : {
            "type" : "string"
          },
          "hreflang" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "deprecation" : {
            "type" : "string"
          },
          "profile" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "templated" : {
            "type" : "boolean"
          }
        }
      },
      "Links" : {
        "type" : "object",
        "additionalProperties" : {
          "$ref" : "#/components/schemas/Link"
        }
      },
      "ErrorResponse" : {
        "description" : "Стандартный формат ошибки API (RFC 9457 Problem Details)",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP статус код",
            "example" : 404
          },
          "type" : {
            "type" : "string",
            "description" : "URI типа ошибки",
            "example" : "https://api.rutcampustrack.ru/problems/resource-not-found"
          },
          "title" : {
            "type" : "string",
            "description" : "Краткое описание ошибки",
            "example" : "Ресурс не найден"
          },
          "detail" : {
            "type" : "string",
            "description" : "Детальное описание",
            "example" : "Группа с id=99 не найдена"
          },
          "instance" : {
            "type" : "string",
            "description" : "URI запроса, вызвавшего ошибку",
            "example" : "/api/academic/groups/99"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Время возникновения ошибки (UTC)",
            "example" : "2026-04-24T10:15:30Z"
          },
          "traceId" : {
            "type" : "string",
            "description" : "Correlation ID из MDC (для трассировки в логах)",
            "example" : "abc-trace-123"
          },
          "fieldErrors" : {
            "type" : "array",
            "description" : "Ошибки валидации полей body DTO (только для 400)",
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            }
          },
          "field" : {
            "type" : "string",
            "description" : "Имя поля DTO, вызвавшего конфликт (только для 409, BUG-006-2)",
            "example" : "login"
          },
          "extras" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Дополнительные данные для клиента (счётчики каскадного удаления, retry-after, и т.п.)",
              "example" : {
                "scheduleItemsCount" : 3
              }
            },
            "description" : "Дополнительные данные для клиента (счётчики каскадного удаления, retry-after, и т.п.)",
            "example" : {
              "scheduleItemsCount" : 3
            }
          }
        }
      },
      "FieldError" : {
        "description" : "Ошибка валидации одного поля DTO (RFC 9457 Extension Member)",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "Имя поля DTO",
            "example" : "displayName"
          },
          "rejectedValue" : {
            "type" : "object",
            "description" : "Отклонённое значение"
          },
          "message" : {
            "type" : "string",
            "description" : "Локализованное сообщение об ошибке",
            "example" : "Имя не может быть пустым"
          }
        }
      }
    }
  }
}